# specs/api.yml
openapi: 3.0.3
info:
  title: MilkRound Customer API
  description: |
    APIs for customers to manage orders and deliveries through the MilkRound application.
    These endpoints are intended for users who are making orders and managing their deliveries.

    ## Authentication

    The onboarding flow is unauthenticated. A `customerId` (UUID) is issued after address
    confirmation via `POST /Customers` and must be passed as the `X-Customer-Id` header on
    all subsequent requests.
  version: 2.0.0
  contact:
    name: MilkRound Support

servers:
  - url: https://api.milkround.com/v1
    description: Production server
  - url: https://api.staging.milkround.com/v1
    description: Staging server

tags:
  - name: Onboarding
    description: Unauthenticated onboarding flow — validate supplier code, check service area, create customer and initial subscription
  - name: Suppliers
    description: Supplier information and schedules
  - name: PlannedDeliveries
    description: Management of specific planned delivery instances
  - name: DeliverySchedules
    description: Management of recurring delivery schedules

paths:

  # ── Onboarding ──────────────────────────────────────────────────────────────

  /Suppliers/ValidateCode:
    # NOTE ON THE `query` OPERATION: this endpoint uses the HTTP QUERY method —
    # a body-bearing, safe, idempotent method (see the IETF draft
    # draft-ietf-httpbis-safe-method-w-body). OpenAPI 3.0.3's Path Item Object
    # only formally enumerates get/put/post/delete/options/head/patch/trace as
    # operation keys, so tooling built strictly against the 3.0.3 schema will
    # not recognise a `query:` key here and may ignore or reject it. We still
    # express it as `query:` below because that is the most accurate
    # representation of the real wire contract, and treat this as a documented
    # judgment call pending either an OpenAPI version with native support or a
    # vendor-extension convention. Do NOT treat the presence of `post:` (if any
    # tooling forces a fallback) as the source of truth — `query` is the actual
    # method the API and client use.
    query:
      tags:
        - Onboarding
      summary: Validate a supplier invite code
      description: |
        Validates the 6-character code from the customer's welcome card.
        Returns basic identifying information for the Supplier business
        record (not an individual delivery round/rota — a Supplier may run
        multiple rota areas) so the app can personalise the onboarding
        journey. No customer record is created at this stage.

        Uses the HTTP QUERY method — see the note above the `query:` key on
        this path item for why this doesn't appear as `post:` or `get:`.
      operationId: validateSupplierCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateCodeRequest'
      responses:
        '200':
          description: Code is valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateCodeResponse'
        '404':
          description: Code not recognised or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /Suppliers/{id}/ServiceArea:
    get:
      tags:
        - Onboarding
      summary: Check whether a postcode is within the supplier's delivery area
      description: |
        Returns whether the given postcode falls on the supplier's route and,
        if so, the expected delivery window (e.g. "Mon–Sat before 7am").
        Postcode validation and geocoding should be performed client-side via
        postcodes.io before calling this endpoint.
      operationId: checkServiceArea
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: postcode
          in: query
          required: true
          description: UK postcode (spaces optional)
          schema:
            type: string
            example: BS8 2AB
      responses:
        '200':
          description: Service area check result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceAreaResponse'
        '400':
          description: Invalid postcode format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Supplier not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /Customers:
    post:
      tags:
        - Onboarding
      summary: Create a new customer
      description: |
        Creates a customer record after the address has been confirmed as
        in-range. The supplier code is re-asserted to prevent orphaned records.
        Returns a `customerId` that the client persists in secure storage and
        sends as `X-Customer-Id` on all subsequent requests.
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCustomerResponse'
        '404':
          description: Supplier code not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: |
            Conflict. Two distinct conflict cases are possible here, both
            using this same status code for consistency, distinguished by
            the `code` field on `ErrorResponse`:
              - `DUPLICATE_ADDRESS` — a customer with this address already
                exists for this supplier.
              - `OUT_OF_SERVICE_AREA` — the address's postcode does not fall
                within any of the supplier's delivery rota areas. Clients
                should normally catch this earlier via
                `GET /Suppliers/{id}/ServiceArea`, but the server re-checks
                on create since the two calls are not atomic.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /Customers/{id}/Subscription:
    post:
      tags:
        - Onboarding
      summary: Create initial subscription
      description: |
        Batch-creates delivery schedules and initial item selections at the
        end of the onboarding flow. One `DeliverySchedule` is created per
        selected day. `defaultBasket` applies to all days; `dayOverrides`
        replaces the default basket for specific days (the "split a day"
        feature from the product picker).
      operationId: createSubscription
      parameters:
        - name: id
          in: path
          required: true
          description: Customer ID returned by POST /Customers
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionRequest'
      responses:
        '201':
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSubscriptionResponse'
        '400':
          description: Invalid request (no days selected, unknown product IDs)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Suppliers ────────────────────────────────────────────────────────────────

  /Supplier/{id}:
    get:
      tags:
        - Suppliers
      summary: Get supplier information
      operationId: getSupplierInfo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Supplier information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierInfo'
        '404':
          description: Supplier not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /Supplier/{id}/DeliverySchedule:
    get:
      tags:
        - Suppliers
      summary: Get supplier delivery schedule
      description: Returns the delivery slots available from this supplier including frequency, day and cutoff times.
      operationId: getSupplierDeliverySchedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Delivery schedule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeliveryScheduleResponse'
        '404':
          description: Supplier not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /Supplier/{id}/Items:
    get:
      tags:
        - Suppliers
      summary: Get supplier items
      description: Returns the items available from this supplier. Optionally filter by delivery schedule or free-text query.
      operationId: getSupplierItems
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: deliveryScheduleId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: query
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierItemsResponse'
        '404':
          description: Supplier not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /Supplier/{id}/PlannedDeliveries:
    get:
      tags:
        - PlannedDeliveries
      summary: Get planned deliveries for supplier
      operationId: getPlannedDeliveries
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Planned deliveries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlannedDeliveriesResponse'
        '404':
          description: Supplier not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Planned Deliveries ───────────────────────────────────────────────────────

  /PlannedDeliveries/Next:
    get:
      tags:
        - PlannedDeliveries
      summary: Get next scheduled delivery
      operationId: getNextDelivery
      responses:
        '200':
          description: Next delivery
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Delivery'
        '404':
          description: No deliveries scheduled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /PlannedDeliveries/{id}/Items:
    post:
      tags:
        - PlannedDeliveries
      summary: Add item to a specific delivery
      operationId: addItemToPlannedDelivery
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddItemRequest'
      responses:
        '200':
          description: Item added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModifyDeliveryResponse'
        '404':
          description: Planned delivery not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /PlannedDeliveries/{id}/Items/{itemId}:
    delete:
      tags:
        - PlannedDeliveries
      summary: Remove item from a specific delivery
      operationId: removeItemFromPlannedDelivery
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Item removed
        '404':
          description: Planned delivery or item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /PlannedDeliveries/{id}/Cancel:
    post:
      tags:
        - PlannedDeliveries
      summary: Skip a planned delivery
      operationId: cancelPlannedDelivery
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Delivery skipped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModifyDeliveryResponse'
        '404':
          description: Planned delivery not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Delivery Schedules ───────────────────────────────────────────────────────

  /DeliverySchedules/{id}/Items:
    post:
      tags:
        - DeliverySchedules
      summary: Add item to a delivery schedule
      operationId: addItemToDeliverySchedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddItemRequest'
      responses:
        '200':
          description: Item added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModifyDeliveryResponse'
        '404':
          description: Delivery schedule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /DeliverySchedules/{id}/Items/{itemId}:
    delete:
      tags:
        - DeliverySchedules
      summary: Remove item from a delivery schedule
      operationId: removeItemFromDeliverySchedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Item removed
        '404':
          description: Delivery schedule or item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /DeliverySchedules/{id}/Cancel:
    post:
      tags:
        - DeliverySchedules
      summary: Cancel a delivery schedule
      operationId: cancelDeliverySchedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Schedule cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModifyDeliveryResponse'
        '404':
          description: Delivery schedule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:

    # ── Onboarding schemas ─────────────────────────────────────────────────────

    ValidateCodeRequest:
      type: object
      required:
        - code
      properties:
        code:
          type: string
          description: 6-character invite code from the customer's welcome card
          example: ABC123

    ValidateCodeResponse:
      type: object
      required:
        - supplierId
        - businessName
      properties:
        supplierId:
          type: string
          format: uuid
          description: |
            Identifies the Supplier record (the business), not an individual
            delivery round/rota. A single supplier may have multiple rota
            areas; this id is scoped to the business as a whole.
        businessName:
          type: string
          example: Adam Cartwright

    ServiceAreaResponse:
      type: object
      required:
        - inRange
      properties:
        inRange:
          type: boolean
        eta:
          type: string
          description: Human-readable delivery window shown on the confirmed-address card
          example: Mon–Sat before 7am

    CustomerAddress:
      type: object
      required:
        - line1
        - line2
        - postcode
      properties:
        flat:
          type: string
          description: Flat or unit number (optional)
          example: Flat 2
        line1:
          type: string
          example: 14 Orchard Lane
        line2:
          type: string
          example: Clifton
        postcode:
          type: string
          example: BS8 2AB
        notes:
          type: string
          description: Free-text delivery instructions (e.g. "Leave in the side porch")
        latitude:
          type: number
          format: double
          description: From postcodes.io — used for route optimisation
        longitude:
          type: number
          format: double

    CreateCustomerRequest:
      type: object
      required:
        - name
        - supplierCode
        - address
      properties:
        name:
          type: string
          description: The customer's name
          example: Jane Wilson
        supplierCode:
          type: string
          description: Re-asserted so the request is self-contained and auditable
          example: ABC123
        address:
          $ref: '#/components/schemas/CustomerAddress'

    CreateCustomerResponse:
      type: object
      required:
        - customerId
      properties:
        customerId:
          type: string
          format: uuid

    BasketItem:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1

    CreateSubscriptionRequest:
      type: object
      required:
        - deliveryDays
        - defaultBasket
      properties:
        deliveryDays:
          type: array
          description: Days of the week to receive deliveries
          minItems: 1
          items:
            $ref: '#/components/schemas/DeliveryDayOfWeek'
          example:
            - Monday
            - Wednesday
            - Friday
        defaultBasket:
          type: array
          description: Items delivered on all selected days (unless overridden)
          items:
            $ref: '#/components/schemas/BasketItem'
        dayOverrides:
          type: object
          description: |
            Per-day basket overrides. Keys are day names matching `deliveryDays`.
            An override replaces the `defaultBasket` entirely for that day.
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/BasketItem'
          example:
            Monday:
              - productId: 00000000-0000-0000-0000-000000000001
                quantity: 4

    CreateSubscriptionResponse:
      type: object
      required:
        - scheduleIds
      properties:
        scheduleIds:
          type: array
          items:
            type: string
            format: uuid
        firstDeliveryDate:
          type: string
          format: date-time
          description: Scheduled date/time of the first delivery

    # ── Shared / existing schemas ──────────────────────────────────────────────

    Price:
      type: object
      required:
        - amount
        - precision
        - symbol
      properties:
        amount:
          type: number
          format: decimal
          example: 2.50
        precision:
          type: integer
          example: 2
        symbol:
          type: string
          example: "£"

    Item:
      type: object
      required:
        - id
        - name
        - price
        - iconName
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Semi-Skimmed Milk 1L
        price:
          $ref: '#/components/schemas/Price'
        iconName:
          type: string
          example: milk-bottle

    SupplierItem:
      allOf:
        - $ref: '#/components/schemas/Item'
        - type: object
          required:
            - maximumQuantity
          properties:
            maximumQuantity:
              type: integer
              minimum: 1
              example: 10

    DeliveryStatus:
      type: string
      enum:
        - Pending
        - Packed
        - Shipped
        - Cancelled
        - Delivered

    Delivery:
      type: object
      required:
        - id
        - items
        - deliverySlotId
        - scheduledTime
        - status
      properties:
        id:
          type: string
          format: uuid
        items:
          type: array
          items:
            $ref: '#/components/schemas/DeliveryItem'
        deliverySlotId:
          type: string
          format: uuid
        scheduledTime:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/DeliveryStatus'

    DeliveryItem:
      type: object
      required:
        - item
        - quantity
      properties:
        item:
          $ref: '#/components/schemas/Item'
        quantity:
          type: integer
          minimum: 1

    DeliveryDayOfWeek:
      type: string
      enum:
        - Monday
        - Tuesday
        - Wednesday
        - Thursday
        - Friday
        - Saturday
        - Sunday

    DeliverySlot:
      type: object
      required:
        - id
        - dayOfWeek
        - frequency
        - cutoffTime
      properties:
        id:
          type: string
          format: uuid
        dayOfWeek:
          $ref: '#/components/schemas/DeliveryDayOfWeek'
        frequency:
          type: string
          enum:
            - Weekly
            - Fortnightly
            - Monthly
        cutoffTime:
          type: string
          format: time
          example: "18:00"
        cutoffDaysBefore:
          type: integer
          minimum: 0
          example: 1

    DeliveryScheduleResponse:
      type: object
      required:
        - slots
      properties:
        slots:
          type: array
          items:
            $ref: '#/components/schemas/DeliverySlot'

    SupplierInfo:
      type: object
      required:
        - id
        - businessName
      properties:
        id:
          type: string
          format: uuid
        businessName:
          type: string
          example: "Smith's Dairy Farm"
        phoneNumber:
          type: string
          example: "+44 1234 567890"
        emailAddress:
          type: string
          format: email

    SupplierItemsResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SupplierItem'

    AddItemRequest:
      type: object
      required:
        - productId
        - deliveryScheduleIds
        - quantity
      properties:
        productId:
          type: string
          format: uuid
        deliveryScheduleIds:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
        quantity:
          type: integer
          minimum: 1

    ModifyDeliveryResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          format: uuid

    PlannedDeliveriesResponse:
      type: object
      required:
        - deliveries
      properties:
        deliveries:
          type: array
          items:
            $ref: '#/components/schemas/Delivery'

    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        code:
          type: string
          description: |
            Machine-readable error code. Distinguishes between conflict kinds
            that share the same HTTP status, e.g. on `POST /Customers`:
            `DUPLICATE_ADDRESS` vs `OUT_OF_SERVICE_AREA` (both 409).